Adds a column with the appropriate grid cells.
However, it only works if the grid is filled. In other words, if there
is a 3 X 2 grid (3 rows, 2 columns), then the maximum
number of cells is 6. The add column feature in this commit assumes that
the grid is filled with it's maximum number of cells.
For our purposes, this should be sufficient since we are automatically
filling the grid whenever we add a column / row (and
we are starting off with a filled grid).

Note that although the CSS property defines the default column and row
template values, this does not carry over to Javascript.
That is, even though we set the property "grid-template-columns: 100px
100px" , the value
document.getElementById("grid-board").style.gridTemplateColumns does not
equal to 100px 100px. It is equal to an empty string.
This is why you'll see variables initialized to the default grid size in
the JavaScript file.

Pseudocode I made and followed:

/*To add a column, append a value (e.g. 100px) to the grid-template-columns
*property. This will add a new column. The number of new elements to add
*should be the number of row elements (use split with " " as a delimeter
*to find the total number of new elements to add (to fill the column)
*/