Properly adds new cells to the new column
Before, new cells would be distributed through all columns.
I foresaw that this would be a problem, since we are planning on having
the user style the color of the cells. Therefore, it is pertinent that
we keep the order of the cells in order to avoid user frustration when
their cells shift by adding a new column.

To remedy this, I modified the fucntion to take in the children elements
and insert new cells by going down the children collection. The decrement
value is taken based on the number of columns (before adding a new
column).

IMPORTANT! The children collection is a live HTML collection, which
means that to iterate over the elements as one would do with an array,
we must decrement from the end of the collection. Incrementing would not
work as we are adding elements, which increases the length of the array
and thus is likely to be problematic.

See the section "Creating Nodes" in this link for a brief introduction
to live collections: https://eloquentjavascript.net/14_dom.html