Added remove column button and functionality.
The remove column functionality removes the right-most column along with
the appropriate cells. The process is similar to the add column feature,
except that we are slicing off a column declaration in the style
property, and instead of adding new cells using a for loop, we are
removing the current cell.

Note the following edge case: What happens if there are no columns to
remove? Answer: The for loop is not entered since i will be less than 0
due to the HTML Collection having a length of 0, and i starts at the
length - 1.

But so does the for loop in the add function. Therefore, we cannot add
columns if all of them are removed. To fix this, an if statement was
added where it adds one cell if there are no cells.