Added a dropdown menu for a user to select a color
Whenever a user selects a color, it is stored in a variable called
selectedColor to be used later.

This is better then just calling the selector element and asking for its
value every time we need it (which we will need it everytime a user
clicks on a box to color it) since we'd need to request the value from
the DOM which costs more time then just having it saved.