Fixed bug with drag & drop, added awesome styling
This commit fixes a bug with the drag and drop where the user can paint
multiple tiles without clicking the mouse. The cause of this bug was
difficult to pin down. Essentially, when a user clicks a cell and tries
to drag it, their cursor would sometimes change to a "Not allowed"
symbol (circle with a slash across it). This would cause the mouse up
event to not register when the user releases their mouse, allowing them
to paint tiles without holding down the mouse key.

Based on several rounds of testing and debugging, it seems that this
occurs because the user is selecting the empty div element contents
(think of dragging and dropping text from a page into the search bar).
The "Not Allowed" symbol appears because the current spot the user is at
is not a valid place to drop the content, but they are allowed to drop
it in a text box.

Understanding this, the bug can be solved by setting the "user-select"
CSS property to the value "none", which prevents a user from selecting
text from elements that have this property. This property was added to
the body element, and it seems to have fixed the bug.

Additionally, the front page was styled using fancy CSS properties. It
is my first time using these properties, but they seem to have the
effect I was going for. Play around with them and check out the MDN docs
to see how they work--they're pretty cool!