Improved user experience via styling
The box shadows for the buttons did not match their linear-gradient (the
shadows should have been casted towards the other side). That is now
fixed by redirecting the shadows.

The alert message that pops up when the maximum number of columns are
reached now disappears over the course of 5 seconds (2 seconds it stays,
then fades over 3 seconds). Once it disappears, the text is set to empty
to allow more room on the page for the grid cells.

The borders for the grid cells have been increased from 1px to 5px,
giving them a more solid shape. The grid cells themselves have had their
sizes decreased to 50px. This allows for users to draw more interesting
shapes with the added boxes.